// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/processAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';

export interface ProcessAC extends BaseModelAC { 
    parentProcessName?: string;
    name?: string;
    entityId: string;
    parentId?: string;
    scopeBasedOn?: string;
    parentProcess?: ProcessAC;
    subProcess?: Array<ProcessAC>;
    auditPlanId?: string;
    scope?: string;
}